Moved entity types in `meta` module.
