--- librad/src/id/user.rs
@@ -1,39 +1,0 @@
-// This file is part of radicle-link
-// <https://github.com/radicle-dev/radicle-link>
-//
-// Copyright (C) 2019-2020 The Radicle Team <dev@radicle.xyz>
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License version 3 or
-// later as published by the Free Software Foundation.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see <https://www.gnu.org/licenses/>.
-
-use crate::id::entity::{data::EntityData, Entity, Error};
-use serde::{Deserialize, Serialize};
-
-#[derive(Clone, Serialize, Deserialize, Default)]
-pub struct UserInfo {
-    pub email: String,
-}
-
-pub type UserData = EntityData<UserInfo>;
-
-impl UserData {
-    pub fn set_email(mut self, email: String) -> Self {
-        self.info.email = email;
-        self
-    }
-
-    pub fn build(self) -> Result<User, Error> {
-        User::from_data(self)
-    }
-}
-
-pub type User = Entity<UserInfo>;
